session 설정 app에서 함
